+++ HISTORY.rst
@@ -12,6 +12,7 @@ History
 * Drop support for adding `validators` as assigment to `Event`.
 * `TransitionNotAllowed` changed internal attr from `transition` to `event`.
 * `StateMachine.run` is deprecated in favor of `StateMachine.send`.
+* `StateMachine.allowed_transitions` is deprecated in favor of `allowed_events`.
 
 
 0.9.0 (2022-12-21)